Split unit tests into separate files
